+++ backend/unimate/src/main/java/com/unimate/domain/user/user/service/UserAuthService.java
@@ -1,8 +1,11 @@
 package com.unimate.domain.user.user.service;
 
 import com.unimate.domain.user.user.dto.UserLoginRequest;
+import com.unimate.domain.user.user.dto.UserSignupRequest;
+import com.unimate.domain.user.user.dto.UserSignupResponse;
 import com.unimate.domain.user.user.entity.User;
 import com.unimate.domain.user.user.repository.UserRepository;
+import com.unimate.domain.verification.service.VerificationService;
 import com.unimate.global.auth.dto.Tokens;
 import com.unimate.global.auth.model.SubjectType;
 import com.unimate.global.auth.service.TokenService;
@@ -19,6 +22,32 @@ public class UserAuthService {
     private final UserRepository userRepository;
     private final BCryptPasswordEncoder passwordEncoder;
     private final TokenService tokenService;
+    private final VerificationService verificationService;
+
+
+    @Transactional
+    public UserSignupResponse signup(UserSignupRequest req) {
+        if (userRepository.existsByEmail(req.getEmail())) {
+            throw ServiceException.badRequest("이미 가입된 이메일입니다.");
+        }
+
+        verificationService.assertVerifiedEmailOrThrow(req.getEmail());
+
+        User user = new User(
+                req.getName(),
+                req.getEmail(),
+                passwordEncoder.encode(req.getPassword()),
+                req.getGender(),
+                req.getBirthDate(),
+                req.getUniversity()
+        );
+        user.verifyStudent();
+
+        userRepository.save(user);
+        verificationService.consumeVerification(req.getEmail());
+
+        return new UserSignupResponse(user.getId(), user.getEmail(), user.getName());
+    }
 
     @Transactional
     public Tokens login(UserLoginRequest req) {
